requirements.txt
O365
pandas
python-dotenv
openpyxl
lxml
beautifulsoup4